added new classes

+++ src/main/java/com/example/OnlineFoodOrderingSystem/entities/Cart.java
@@ -20,7 +20,7 @@ public class Cart {
     @OneToOne
     private User user;
 
-    private Long total;
+    private int total;
 
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL,orphanRemoval = true)
     private List<CartItem> cartItems = new ArrayList<>();
